requirements.txt
streamlit==1.28.0
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
pandas==2.0.3
python-dotenv==1.0.0
lxml==4.9.3